feat(calendar-ext): 予定管理ページの行ボタンも 3 択 modal に対応 (Refs #229)
- RecurringScopeModal を src/entities/event/ に切り出して共有化
- EventManagement の行「予定化解除 / 予定化する」も recurring instance では
  3 択 modal を経由するように変更 (EventDetailPanel と挙動を統一)
- AppShell から setRecurringEventVisibilityOverride を EventManagement にも渡す

予定タブの一覧から系列操作する経路がこれまで欠けていた (single 操作だけが直接
発火していた)。EventDetailPanel と同じ semantics に揃え、recurring 系列の操作
は常に 3 択 modal を経由する形にする。
